+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Changelog starting from v0.5 (2018/09/18)
 
 - **speed up** improvements for the history display (more than 10x).
 - **speed up** improvements to the CSV and TXT download of history logs (30%).
+- deposits: we don't mention anymore the type of deposits that are sent by us to the exterior world.
 
 ### Upgrade instructions